Tuned up UI after running other Repos.
graph sits correctly, feels more polished.
