+++ go.mod
@@ -4,7 +4,7 @@ go 1.18
 
 require (
 	github.com/gdamore/tcell/v2 v2.4.1-0.20210905002822-f057f0a857a1
-	github.com/pb33f/libopenapi v0.4.11
+	github.com/pb33f/libopenapi v0.4.16
 	github.com/pterm/pterm v0.12.51
 	github.com/rivo/tview v0.0.0-20221117065207-09f052e6ca98
 	github.com/spf13/cobra v1.6.1
