+++ html-report/ui/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en-US">
 <head lang="en">
-    <title>Hello React</title>
+    <title>OpenAPI Changes</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
     <style>
         .preloader {
