+++ html-report/ui/src/components/report-statistics/ReportSummary.tsx
@@ -8,36 +8,43 @@ export interface ReportSummaryProps {
     changeStats: ChangeStatistics
 }
 
-export const ReportSummary = (props: ReportSummaryProps) => (
-    <section className='report-summary'>
-        <section className="report-summary-stats">
-            <Statistic title="Added" className="width-100 m-bottom-20"
-                       value={props.changeStats.added} prefix={<PlusSquareOutlined/>}
-            />
-            <Statistic title="Removed" className="width-100 m-bottom-20"
-                       value={props.changeStats.removed} prefix={<MinusSquareOutlined/>}
-            />
-            <Statistic title="Modified" className="width-100 m-bottom-20"
-                       value={props.changeStats.modified} prefix={<EditOutlined/>}
-            />
+export function ReportSummary(props: ReportSummaryProps) {
+    let breakingStat: JSX.Element | undefined;
+    if (props.changeStats.totalBreaking > 0) {
+        breakingStat =  <Statistic title="Breaking" className="width-100 m-bottom-20"
+                                   value={props.changeStats.totalBreaking}
+                                   valueStyle={{fontWeight: 'bolder', color: 'var(--error-font-color)'}}/>
+    }
+    return (
+        <section className='report-summary'>
+            <section className="report-summary-stats">
+                <Statistic title="Added" className="width-100 m-bottom-20"
+                           value={props.changeStats.added} prefix={<PlusSquareOutlined/>}
+                />
+                <Statistic title="Removed" className="width-100 m-bottom-20"
+                           value={props.changeStats.removed} prefix={<MinusSquareOutlined/>}
+                />
+                <Statistic title="Modified" className="width-100 m-bottom-20"
+                           value={props.changeStats.modified} prefix={<EditOutlined/>}
+                />
 
-            <Statistic title="Total" className="width-100 m-bottom-20"
-                       value={props.changeStats.total}
-            />
-            <Statistic title="Breaking" className="width-100 m-bottom-20"
-                       value={props.changeStats.totalBreaking}
-                       valueStyle={{fontWeight: 'bolder', color: 'var(--error-font-color)'}}/>
-        </section>
-        <section className='commit'>
+                <Statistic title="Total" className="width-100 m-bottom-20"
+                           value={props.changeStats.total}
+                />
+                {breakingStat}
+            </section>
+            <section className='commit'>
             <span className='commit-changed'>
-                Commit: {props.changeStats.commit.hash.substring(0, 6)} | {props.changeStats.commit.date} by&nbsp;
+                Commit: {props.changeStats.commit.hash.substring(0, 6)} |&nbsp;
+                <span className='commit-date'>{props.changeStats.commit.date}</span> by&nbsp;
                 <span className='commit-author'>
                     <a href={`mailto:${props.changeStats.commit.authorEmail}`}>{props.changeStats.commit.author}</a>
                </span>
             </span>
 
-                 {props.changeStats.commit.message}
+                {props.changeStats.commit.message}
 
+            </section>
         </section>
-    </section>
-);
+    )
+};
