+++ html-report/ui/src/variables.css
@@ -72,6 +72,7 @@
     --hrcolor: #3d3d3d;
     --code-border: #525252;
     --string-literal: #fd66e5;
+    --neon-pink: #E400FB;
     --warn-400: #ff6a00;
     --warn-400-lowalpha: rgba(255, 106, 0, 0.15);
     --warn-300: #ff9900;
